+++ Outils_Decision/TP1/Solution.cpp
@@ -1,7 +1,7 @@
 #include "Solution.hpp"
 
 // Methods
-void Solution::insertM(int i, int val)
+void Solution::insertM(int i, float val)
 {
 	m[i] = val;
 }
@@ -16,7 +16,7 @@ void Solution::insertNiv(int i, int val)
 	niv[i] = val;
 }
 
-int Solution::getM(int i)
+float Solution::getM(int i)
 {
 	return m[i];
 }
